Спринт 12 - до удаления UserMemoryStorage

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,33 +1,24 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.controller.UserController;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
+import ru.yandex.practicum.filmorate.validator.UserValidator;
 
-import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserController.class);
-
     private final UserStorage userStorage;
 
-    private static final String MESSAGE_OF_VALID_EMAIL_EMPTY = "Электронная почта не может быть пустой";
-    private static final String MESSAGE_OF_VALID_EMAIL_SYMBOL = "Символ @ - не передан";
-    private static final String MESSAGE_OF_VALID_LOGIN_EMPTY = "Логин не может быть пустой";
-    private static final String MESSAGE_OF_VALID_LOGIN_BLANK = "Логин не может состоять из пробелов";
-    private static final String MESSAGE_OF_VALID_LOGIN_SPACE = "Логин не может содержать пробелов";
-    private static final String MESSAGE_OF_VALID_BIRTHDAY = "Дата рождения не может быть в будущем";
     private static final String MESSAGE_OF_ID_USER = "Некорректный Id пользователя";
     private static final String MESSAGE_OF_NULL_ID_USER = "Id пользователя не указан";
     private static final String MESSAGE_OF_NOT_FOUND_USER = "Пользователь не найден";
@@ -35,10 +26,12 @@ public class UserService {
 
 
     public Collection<User> findAll() {
+        log.info("Запрос на получение списка пользователей");
         return userStorage.findAll();
     }
 
     public User findById(long id) {
+        log.info("Запрос на получение пользователя по id = {}", id);
         if (id <= 0) {
             throw new ValidationException(MESSAGE_OF_ID_USER);
         }
@@ -49,11 +42,13 @@ public class UserService {
     }
 
     public User create(User user) {
-        validateUser(user);
+        log.info("Запрос создания пользователя");
+        UserValidator.validateUser(user);
         return userStorage.create(user);
     }
 
     public User update(User user) {
+        log.info("Запрос на обновление информации о пользователе: {}", user);
         if (user.getId() == null) {
             log.debug("ID не указан при обновлении");
             throw new ValidationException(MESSAGE_OF_NULL_ID_USER);
@@ -65,15 +60,26 @@ public class UserService {
         if (!userStorage.containUser(user.getId())) {
             throw new NotFoundException("Пользователь с id = " + user.getId() + " не найден");
         }
-        validateUser(user);
+        UserValidator.validateUser(user);
         return userStorage.update(user);
     }
 
     public void deleteAll() {
-        userStorage.deleteAll();
+        log.info("Удаление всех пользователей");
+        if (!findAll().isEmpty()) {
+            userStorage.deleteAll();
+        }
     }
 
-    public void addFriend(long userId, long idFriend) {
+    public List<User> getFriendList(Long userId) {
+        log.info("Запрос на получение списка друзей, пользователя с Id = {}", userId);
+        findById(userId);
+        return userStorage.getFriendList(userId);
+    }
+
+
+    public void addFriend(long userId, long friendId) {
+        log.info("Запрос на дружбу между {} и {}", userId, friendId);
         validateFriend(userId, idFriend);
 
         userStorage.findById(userId).addFriend(idFriend);
@@ -119,37 +125,6 @@ public class UserService {
         return commons;
     }
 
-    private void validateUser(User user) {
-        if (user.getEmail() == null || user.getEmail().isEmpty() || user.getEmail().isBlank()) {
-            log.error("Передано пустое значение эл. почты - {}", user.getEmail());
-            throw new ValidationException(MESSAGE_OF_VALID_EMAIL_EMPTY);
-        }
-        if (!user.getEmail().contains("@")) {
-            log.error("В наименование email отсутствует символ @ - {}", user.getEmail());
-            throw new ValidationException(MESSAGE_OF_VALID_EMAIL_SYMBOL);
-        }
-
-        if (user.getLogin() == null || user.getLogin().isEmpty()) {
-            log.error("Логин пустой - {}", user.getLogin());
-            throw new ValidationException(MESSAGE_OF_VALID_LOGIN_EMPTY);
-        }
-        if (user.getLogin().isBlank()) {
-            log.error("Логин состоит из пробелов - {}", user.getLogin());
-            throw new ValidationException(MESSAGE_OF_VALID_LOGIN_BLANK);
-        }
-        if (user.getLogin().contains(" ")) {
-            log.error("Логин содержит пробелы - {}", user.getLogin());
-            throw new ValidationException(MESSAGE_OF_VALID_LOGIN_SPACE);
-        }
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.error("Дата рождения находится в будущем - {}", user.getBirthday());
-            throw new ValidationException(MESSAGE_OF_VALID_BIRTHDAY);
-        }
-        if (user.getName() == null || user.getName().isEmpty() || user.getName().isBlank()) {
-            log.debug("Имя пользователя пустое, будет использован логин - {}", user.getLogin());
-            user.setName(user.getLogin());
-        }
-    }
 
     private void validateFriend(Long userId, Long idFriend) {
         if (!userStorage.containUser(userId)) {
